renamed footer components and broke out upgrade view to its own component
