+++ Client/src/components/game/FootArmyView.css
